collateral manager mapping
